+++ src/modules/pos/validator/components/StakeSummary/styles.css
